make sure that _decode() can be called more than once
make sure that _decode() can be re-used on the same
descriptor. what this will effectively do however is to
free the previous rows and patterns in the descriptor
on each _decode() call. so if a song was previously decoded
with a descriptor, decoding a new one should erase the
old song rows and pattern info from that same descriptor.

also fix a mem-leak related to 'nrows'.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>