fix potential memory leaks
when a segment is appended to another segment
realloc() is used, thus the result is a new heap location.
_free_segment() should be used to release such locations, but
_alloc_x_segment did not allocate / copy the memory from the
main buffer instead it only referenced it.

to solve the issue we now free both the segment pointer itself
and the data in _free_segment, but also when a new segment
is allocated we make a copy of the data from the main decoded
buffer.

more memory but safer operation.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>